requirements.txt
alpaca-py==0.28.0
annotated-types==0.7.0
asttokens==2.4.1
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cuda-python==12.5.0
cudf-cu12==24.6.1
cupy-cuda12x==13.2.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.0.1
fastrlock==0.8.2
fonttools==4.53.1
frozendict==2.4.4
fsspec==2024.6.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
llvmlite==0.43.0
lxml==5.2.2
markdown-it-py==3.0.0
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.0.8
multitasking==0.0.11
nest-asyncio==1.6.0
numba==0.60.0
numpy==1.26.4
nvtx==0.2.10
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.6
pendulum==3.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pynvjitlink-cu12==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
requests==2.32.3
rich==13.7.1
rmm-cu12==24.6.0
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
sseclient-py==1.8.0
stack-data==0.6.3
tabulate==0.9.0
time-machine==2.14.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yfinance==0.2.41
